+++ src/cointoss/domain/model/coin_toss_result.py
