+++ src/cointoss/presentation/schema/cointoss.py
