+++ src/minigame/domain/model/minigame.py
