+++ src/minigame/presentation/schema/minigame.py
