+++ src/plinko/domain/model/plinko_result.py
