+++ src/ticket/domain/model/ticket.py
