+++ src/ticket/presentation/schema/ticket.py
