+++ src/yavarwee/domain/model/yavarwee_result.py
